Connection TourLog Backend Changes

+++ backend/src/main/java/at/technikum/backend/service/TourLogService.java
@@ -4,8 +4,11 @@ import at.technikum.backend.dto.request.RequestTourLogDto;
 import at.technikum.backend.dto.response.ResponseTourLogDto;
 import at.technikum.backend.entity.Tour;
 import at.technikum.backend.entity.TourLog;
+import at.technikum.backend.entity.User;
+import at.technikum.backend.mapper.TourLogMapper;
 import at.technikum.backend.repository.TourLogRepository;
 import at.technikum.backend.repository.TourRepository;
+import at.technikum.backend.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,12 +24,15 @@ public class TourLogService {
     private final TourLogRepository tourLogRepository;
     private final TourLogMapper mapper;
     private final TourRepository tourRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public ResponseTourLogDto save(UUID tourId, RequestTourLogDto requestTourLogDto) {
+    public ResponseTourLogDto save(UUID tourId, RequestTourLogDto requestTourLogDto, String username) {
         Tour tour = tourRepository.findById(tourId).orElseThrow(() -> new RuntimeException("Tour not found"));
+        User currentUser = userRepository.findByUsername(username).orElseThrow(() -> new EntityNotFoundException("User not found"));
         TourLog entity = mapper.toEntity(requestTourLogDto);
         entity.setTour(tour);
+        entity.setAuthor(currentUser);
         TourLog savedEntity = tourLogRepository.save(entity);
         return mapper.toResponseDto(savedEntity);
     }
@@ -42,15 +48,19 @@ public class TourLogService {
 
     @Transactional(readOnly = true)
     public ResponseTourLogDto getById(UUID tourId, UUID tourLogId) {
-        TourLog tourLog = findAndValidateLog(tourLogId, tourId);
+        TourLog tourLog = findAndValidateLog(tourId, tourLogId);
         return mapper.toResponseDto(tourLog);
     }
 
     @Transactional
-    public ResponseTourLogDto update(UUID tourId, UUID tourLogId, RequestTourLogDto requestTourLogDto) {
+    public ResponseTourLogDto update(UUID tourId, UUID tourLogId, RequestTourLogDto requestTourLogDto, String username) {
         TourLog existingTourLog = findAndValidateLog(tourId, tourLogId);
 
-        existingTourLog.setAuthor(requestTourLogDto.getAuthor());
+        //darf der Nutze überhaupt bearbeiten?
+        if(!existingTourLog.getAuthor().getUsername().equals(username)) {
+            throw new IllegalArgumentException("User not allowed to update this Log!");
+        }
+        //existingTourLog.setAuthor(requestTourLogDto.getAuthor());
         existingTourLog.setDate(requestTourLogDto.getDate());
         existingTourLog.setTime(requestTourLogDto.getTime());
         existingTourLog.setRating(requestTourLogDto.getRating());
@@ -73,8 +83,11 @@ public class TourLogService {
     }
 
     @Transactional
-    public ResponseTourLogDto delete(UUID tourId, UUID tourLogId) {
+    public ResponseTourLogDto delete(UUID tourId, UUID tourLogId, String username) {
         TourLog tourLog = findAndValidateLog(tourId, tourLogId);
+        if(!tourLog.getAuthor().getUsername().equals(username)) {
+            throw new IllegalArgumentException("User not allowed to delete this Log!");
+        }
         tourLogRepository.delete(tourLog);
         return mapper.toResponseDto(tourLog);
     }
